Refactor RewardCell to RewardCardView
